backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
google-generativeai
openai-whisper
moviepy
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
youtube-transcript-api
yt-dlp
sqlalchemy
